+++ lib/arc-furnace.rb
@@ -19,6 +19,7 @@ require 'arc-furnace/inner_join'
 require 'arc-furnace/merging_hash'
 require 'arc-furnace/multi_csv_source'
 require 'arc-furnace/outer_join'
+require 'arc-furnace/pipeline'
 require 'arc-furnace/sink'
 require 'arc-furnace/source'
 require 'arc-furnace/suffixed_fixed_column_csv_sink'
