+++ lib/arc-furnace/fixed_column_csv_sink.rb
@@ -5,7 +5,7 @@ module ArcFurnace
     private_attr_reader :fields, :csv
 
     # Expects filename to a filename to open the csv
-    # Expects fields to a hash of Column name => column Width
+    # Expects fields to a hash of Column name => column count
     def initialize(filename: , fields: , encoding: 'UTF-8', force_quotes: false)
       @csv = CSV.open(filename, 'wb', encoding: encoding, headers: true, force_quotes: force_quotes)
       @fields = fields
